+++ src/extraction/special/refs/indirect.rs
@@ -5,7 +5,10 @@ use winnow::{
     BStr, PResult, Parser,
 };
 
-use crate::extraction::{extract, Extract};
+use crate::{
+    extraction::{extract, Extract},
+    follow_refs::{Build, BuilderParser},
+};
 
 use super::ReferenceId;
 
@@ -27,7 +30,10 @@ impl<'de, T> Indirect<T> {
     /// Allows more generic extraction method by supplying a dedicated parser.
     ///
     /// This allows Livre to have `Indirect` be [`Extract`], and still support an ersatz of
-    /// [`Build`](crate::builder::Build).
+    /// [`Build`](crate::follow_refs::Build).
+    ///
+    /// We go the extra mile and extract the trailing `endobj` tag. This is not actually needed,
+    /// although it does serve as a kind of sanity check.
     pub fn parse<P>(input: &mut &'de BStr, parser: P) -> PResult<Self>
     where
         P: Parser<&'de BStr, T, ContextError>,
@@ -44,8 +50,6 @@ impl<'de, T> Indirect<T> {
     }
 }
 
-/// We go the extra mile and extract the trailing `endobj` tag. This is not actually needed,
-/// although it does serve as a kind of sanity check.
 impl<'de, T> Extract<'de> for Indirect<T>
 where
     T: Extract<'de>,
@@ -55,11 +59,23 @@ where
     }
 }
 
+impl<'de, T> Build<'de> for Indirect<T>
+where
+    T: Build<'de>,
+{
+    fn build<B>(input: &mut &'de BStr, builder: &B) -> PResult<Self>
+    where
+        B: crate::follow_refs::Builder<'de>,
+    {
+        Self::parse(input, builder.as_parser())
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::fmt::Debug;
 
-    use crate::builder::BuilderParser;
+    use crate::follow_refs::BuilderParser;
 
     use super::*;
     use rstest::rstest;
@@ -70,7 +86,7 @@ mod tests {
     #[case(b"0 0 obj\n10    true\nendobj", Indirect{id: (0, 0).into(), inner: (10i32, true)})]
     fn extraction_and_build<'de, T>(#[case] input: &'de [u8], #[case] expected: Indirect<T>)
     where
-        T: Extract<'de> + Debug + PartialEq,
+        T: Extract<'de> + Build<'de> + Debug + PartialEq,
     {
         let result = extract(&mut input.as_ref()).unwrap();
         assert_eq!(expected, result);