+++ src/pdf/mod.rs
@@ -42,8 +42,8 @@
 //! 2. Iterate through the Pages dictionary.
 
 mod content;
-mod pages;
+// mod pages;
 mod trailer_block;
 
-pub use pages::Pages;
+// pub use pages::Pages;
 pub use trailer_block::{RefLocation, StartXRef, Trailer, XRefTrailerBlock};